fix(entities): correct ENUM columnDefinition and Short/Integer type mappings
Add columnDefinition to all ENUM-typed String fields so Hibernate does not
mismatch CHAR vs VARCHAR on schema inspection. Change SMALLINT UNSIGNED columns
(ram_total_gb, ram_velocidad_mhz, tiempo_respuesta_min, tiempo_resolucion_min)
from Integer to Short to match the actual SQL type; add .shortValue()/.intValue()
adapters in CatalogoService and EquipoService, and update affected unit tests.

+++ backend/src/main/java/pe/edu/emch/sgi/service/CatalogoService.java
@@ -187,8 +187,8 @@ public class CatalogoService {
     public TipoIncidenteResponse configurarSla(Integer idTipo, SlaConfigRequest request) {
         TipoIncidente tipo = tipoIncidenteRepository.findById(idTipo)
             .orElseThrow(() -> new ResourceNotFoundException("Tipo de incidente no encontrado: " + idTipo));
-        tipo.setTiempoRespuestaMin(request.getTiempoRespuestaMin());
-        tipo.setTiempoResolucionMin(request.getTiempoResolucionMin());
+        tipo.setTiempoRespuestaMin(request.getTiempoRespuestaMin().shortValue());
+        tipo.setTiempoResolucionMin(request.getTiempoResolucionMin().shortValue());
         return toTipoIncidenteResponse(tipoIncidenteRepository.save(tipo));
     }
 
@@ -242,8 +242,8 @@ public class CatalogoService {
         TipoIncidenteResponse r = new TipoIncidenteResponse();
         r.setIdTipoIncidente(t.getIdTipoIncidente());
         r.setNombreTipo(t.getNombreTipo());
-        r.setTiempoRespuestaMin(t.getTiempoRespuestaMin());
-        r.setTiempoResolucionMin(t.getTiempoResolucionMin());
+        r.setTiempoRespuestaMin(t.getTiempoRespuestaMin().intValue());
+        r.setTiempoResolucionMin(t.getTiempoResolucionMin().intValue());
         r.setDescripcion(t.getDescripcion());
         return r;
     }
